+++ packages/popmotion/src/input/pointer/index.ts
@@ -7,7 +7,7 @@ import { Point2D, PointerProps } from './types';
 const getFirstTouch = ([firstTouch]: Point2D[]): Point2D => firstTouch;
 
 const pointer = (props: PointerProps = {}): Action => getIsTouchDevice()
-  ? touch(props).pipe(getFirstTouch)
+  ? touch(props).pipe(({ touches }) => touches, getFirstTouch)
   : mouse(props);
 
 export default ({ x, y, ...props }: PointerProps = {}): Action => {