+++ site/templates/Homepage/USPs/index.js
@@ -10,22 +10,6 @@ const tweenCode = `tween({
   ease: easing.backOut
 })`;
 
-const springCode = `const ball = document.getElementById('ball');
-const ballStyler = styler(ball);
-
-const ballXY = value({ x: 0, y: 0});
-const ballXY.subscribe(ballStyler.set);
-
-listen(ball, 'mousedown touchstart').start(() => {
-  pointer(ballXY.get())
-    .start(ballXY);
-});
-
-listen(document, 'mouseup touchend').start(() => {
-  spring({ velocity: ballXY.getVelocity() })
-    .start(ballXY);
-});`;
-
 export default () => (
   <Container>
     <Blurb>
@@ -36,6 +20,62 @@ export default () => (
         title="Tween"
         link="/api/tween"
         description="Animate between two states with a scrubbable playhead. Includes a full suite of easing, and methods to generate your own."
+      />
+      <Example
+        title="Spring"
+        link="/api/spring"
+        description="An accurate spring simulation that offers smooth animation and a great variety of spring-feels."
+        code={tweenCode}
+      />
+      <Example
+        title="Keyframes"
+        link="/api/keyframes"
+        description="Animate through multiple sequential states with all the power of a single tween."
+        code={tweenCode}
+      />
+      <Example
+        title="Physics"
+        link="/api/physics"
+        description="A lightweight integrated physics simulation that can be affected mid-animation."
+        code={tweenCode}
+      />
+      <Example
+        title="Decay"
+        link="/api/decay"
+        description="Exponentially-decayed velocity to smoothly dampen momentum."
+        code={tweenCode}
+      />
+    </ExampleSection>
+    <ExampleSection title="Input">
+      <Example
+        title="Pointer"
+        link="/api/pointer"
+        description="Provides absolute or relative pointer coordinates."
+        code={tweenCode}
+      />
+      <Example
+        title="Mulittouch"
+        link="/api/touch"
+        description=""
+        code={tweenCode}
+      />
+      <Example
+        title="Listen"
+        link="/api/listen"
+        description=""
+        code={tweenCode}
+      />
+    </ExampleSection>
+    <ExampleSection title="More ">
+      <Example
+        title="Stagger"
+        description=""
+        code={tweenCode}
+      />
+      <Example
+        title="Animate colors"
+        description="Popmotion uses advanced color blending that avoids muddy saturation loss."
+        code={tweenCode}
       />
     </ExampleSection>
   </Container>