+++ packages/popmotion/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Popmotion adheres to [Semantic Versioning](http://semver.org/).
 
+## [8.0.21@beta] 2017-11-16
+
+### Added
+- `scale` and `rotate` properties to `touch`.
+- `listen` action.
+
 ## [8.0.20@beta] 2017-11-14
 
 ### Fixed